tests/kms_plane: Disable GEN12 media compression YUV tests
The Vebox backend needed for this is not ready yet for YUV formats,
since that needs a separate AUX pagetable setup for the Y and UV planes
and also setting up the matching format in the Vebox tiling convert
command surface state.

It makes sense to test the MC kernel patchset for regressions though, so
let's disable the MC YUV tests until Vebox is able to handle those
formats (and revert this patch when that happens and before the kernel
patches enabling MC YUV formats get merged).

Cc: Mika Kahola <mika.kahola@intel.com>
Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>