tests/kms_ccs: Work around CRC mismatch when mixing SDR/HDR planes
When using both an SDR and HDR plane, the CRC of the resulting image
will not match the CRC of the reference image where just an HDR plane is
used. To avoid this - most probably HW precision issue - make sure all
the planes we enable are of the same SDR/HDR type.

Cc: Mika Kahola <mika.kahola@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>