i915/i915_pm_rc6_residency: Relax constraints for slower HW rc6
Prior to execlists, we are not able to quickly detect the end of the
execution and so rely entirely on HW rc6. This means that each pulse
incurs an extra EI (160ms) before RC6 is entered, dramatically reducing
the amount of time spent in RC6 versus our soft-rc6. As such we need to
increase the tolerance to be much more lenient to allow gen6/gen7 to
pass -- the extreme condition we are trying to prevent happening again
is where we never enter rc6 under these light conditions.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>