intel-ci: Drop gem_ctx_switch from fast feedback
Only a couple of tests from gem_ctx_switch are run in BAT, to check we
have multiple contexts on RCS. It doesn't actually verify the switch,
just that the execbuf API accepts the context argument.

This test is redundant as actual context switching (and more) is verified by
live_gem_contexts and live_gt_contexts selftests.

Instead of using the mediocre gem_ctx_switch stress test in BAT, use
gem_exec_parallel/contexs and gem_exec_parallel/fds as both ensure
multiple contexts are supported for simultaneous use, in a fraction of
the time.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>