intel-ci: Reduce variety of gem_sync in BAT
Historically, we've had many problems with missed interrupt/seqno
syndrome and so have focus on testing with gem_sync. However, these
tests rely on the kernel itself reporting the issue which it no longer
does. So why the extra variety may impose different timing of execution
on the HW (and so different interrupt timings which may or may not help
uncover issues), they do not have any variety in driver coverage. Reduce
the variety (halving the associated runtime) as they are no more likely
to spot an issue than multiple runs through BAT.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>