lib/igt_core: Detect gdb harder
The current gdb detection only works for the parent igt process,
but none of its children will see the gdb and so won't trap properly
in igt_fail_assert(). Also we will not detect gdb if it was attached
after the fact. Fix all of that by looking for the "TracerPid"
information in /proc/<pid>/status. We'll leave the current "assume
parent may be gdb" approach as a fallback.

Also annoyingly by default gdb will only track a single process.
To make it track all of them, and let them all run simultanously
one needs the following incantations:
 set detach-on-fork off
 set schedule-multiple on

Maybe that will save someone from having to trawl as many
docs/gogole hits as I did.

v2: s/1024/32/ bytes (should be more than enough to find the
    TracePid line)
    Put the multi-process gdb instructions into a comment (Chris)

Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>