tests/dumb_buffer: Try to compute the largest possible dumb buffer
For our threaded clear test, we want to create as many buffers as can
fit into the available memory. However, since we don't know the size of
that available memory, it can be easy to create too large or too many
buffers, and waste our time testing the unobtainable. Instead, try
probing the available space by trying to create the largest dumb buffer
we possibly can, and assume that is the equivalent to the available
space.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Ramalingam C <ramalingam.c@intel.com>
Reviewed-by: Ramalingam C <ramalingam.c@intel.com>