tests/i915/gem_mmap_wc: Skip gtt test if aperture is missing
Skip subtest on platforms that do not have mappable aperture.

v2: Skip the test_write_gtt as well when there is no aperture.
It seems like the purpose of this test was to mmap the buffer in
gtt even though it uses the incorrect mapping call. Until this
can be fixed, might be better to skip it. Also remove all subtests
that use run_without_prefault, since this debugfs is being removed.

Signed-off-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Cc: Antonio Argenziano <antonio.argenziano@intel.com>
Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>