tests/kms_fbcon_fbt: Drop master before restoring fbcon
fbdev helpers refuse to restore the fbdev screen if there's still a
master present. This is to avoid fbcon popping up when a compositor
has intentionally disabled all outputs.

Unfortunately the kernel had some bugs in this area, and kms_fbcon_fbt
relied on those. Result was that fbdev restore was skipped, which
skipped the frontbuffer flushing, which broke the PSR subtest's
expectation that PSR is always disable when fbdev is enabled. FBC did
not get broken because FBC was never enabled on linear framebuffers as
used by fbdev.

Fix things up to make sure fbcon gets restored correctly in all cases.

Cc: Noralf Trønnes <noralf@tronnes.org>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>