runner: Support dynamic subtests in testlists
In a very rudimentary and undocumented manner, testlist files can now
have dynamic subtests specified. This feature is intended for very
special cases, and the main supported mode of operation with testlist
files is still the CI-style "run it all no matter what".

The syntax for testlist files is:

igt@binary@subtestname@dynamicsubtestname

As dynamic subtests are not easily listable, any helpers for
generating such testlists are not implemented.

If running in multiple-mode, subtests with dynamic subtests specified
will run in single-mode instead.

Closes: https://gitlab.freedesktop.org/drm/igt-gpu-tools/issues/45
Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>