i915/gem_tiled_.*blits: skip the tests on newer gens
Skip the tests on newer gens when no fences are available
and gem_set_tiling() cannot succeed.

Note that the primary purpose of these tests is to ensure that the
tiling remains consistent across movement of the buffers in the GTT due
to aperture thrashing. As such we would be much better served by a small
selftest to verify position-independent tiling of both physical and
virtual addressing (and where the HW doesn't have such
position-independent tiling, we check we either fixup the swizzling
across relocations, or prevent such relocations.)

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>