i915/gem_set_tiling_vs_blt: Add check for HW tiling support
This test is designed to exercise races between adjusting the tiling of
the object (and so the associated HW fence register) with on-going GPU
execution. (If the GPU depends on the fence register for its detiling,
we cannot update that register until the GPU is idle.) On platforms
without any fences, there can be no race and so nothing to test. (In
fact on any platform after gen4+ the blitter does not use a fence for
itself and so the test has been superfluous.)

Signed-off-by: Ayaz A Siddiqui <ayaz.siddiqui@intel.com>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>