i915/gem_render_tiled_blits: Add check for hardware based FENCE support
This test is verifying the integrity of blitter render copy on platforms
which support FENCE based tiling. Since blitter does not use FENCE after
Gen4+ so this test can only be useful for Platform prior to Gen4.
A check has been added to skip this test if fences  are not available in
platform.

Test to check effects of different physical address values on render
copy is being developed as Kernel selftest.

Signed-off-by: Ayaz A Siddiqui <ayaz.siddiqui@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>