tests/kms_frontbuffer_tracking: Improve tiling test coverage
Lets add some more tiling tests, if tiling is supported by FBC draw
some rectangles and compare the CRC against benchmark if not
supported run the test to guarantee that FBC is disabled.

This is a preparation for when kernel will allow FBC to be enabled
without fences, so we can better test linear and Y tiled
frontbuffers.

v2:
- renamed tile to tiling (Ville)
- removed *-tiling-x test (Ville)
- improved tiling_disable_fbc_subtest documentation (Ville)

v3:
- added new tests description
- fixed failure in prepare_subtest()

v5:
- not adding tiling Y tests to GENs older than 9 because it is not
supported

v6:
- fixed GEN check above

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>