i915/gem_exec_flush: Drop assertion the object is not moved
Each set of relocations track the content of their particular portion of
the batch, the presumed offset they use encode matches their own view.
It is legal for the object to be moved, and the execbuf will have to
relocation -- we can't just assert that the relocations were not
required as that is beyond our own control (unless we switch to
softpin).

Closes: https://gitlab.freedesktop.org/drm/intel/issues/1097
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>