i915/pm_backlight: ensure driver is loaded before checking backlight
Checking the existence of the backlight interface before ensuring the
driver is loaded skips whenever the test is run without the driver
loaded. None of the tests would work properly if they were run as the
first thing or after tests that unload the driver.

Open the driver (and thus probe it as needed) as the first thing.

Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>