intel-ci: Drop b/c pipes from kms_busy fast feedback
The principle test for kms_busy is checking the synchronisation between
on-going rendering to a framebuffer and its flip, that is independent of
the pipe. As such for our fast feedback on driver health, we can look at
the first pipe and assume any errors on the rest will be picked up later
in the shards/idle runs.

Each pass of kms_busy is about 25s.

v2: Use the first pipe connected; combine flip/modeset testing into a
single dynamic group.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Acked-by: Ville Syrjälä <ville.syrjala@linux.intel.com>