tests/kms_rotation_crc: Speed up by not testing every format
As we did with kms_plane_scaling let's not test every pixel
format for rotation fails either. Instead we test each "class"
of formats as defined by igt_reduce_format(). The hope being
that there are no specific bugs in the hw for different
swizzles of the same underlying base format.

On KBL:
$ time ./build/tests/kms_rotation_crc --r sprite-rotation-90
- real	0m39,851s
- user	0m1,037s
- sys	0m12,895s
+ real	0m16,773s
+ user	0m0,357s
+ sys	0m5,475s

As per the usual recipe we add an --extended cmdline knob to
force the full coverage.

Only cuts ~30% of the total kms_rotation_crc runtime though.
There are other subtests which take over 80s on this particular
machine, so further work is clearly required to get the testing
time to sane levels.

Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Martin Peres <martin.peres@linux.intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>