i915: Drop gem_exec_reuse
The test throws a large number of objects at the GPU across many
batches. It only serves to try and assess the scaling impact, without
doing any conformance checking, nor analysing said impact of scaling.
The only effective coverage it gives us is on multi-engine reuse, which
any of the stress tests (gem_exec_parallel, gem_exec_whisper) provide,
and they do validate!

I have no doubt that if we are presented with a concrete problem, or
bug, in this area we will be motivated to write a more precise test
case.

Closes: https://gitlab.freedesktop.org/drm/intel/issues/1289
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Martin Peres <martin.peres@linux.intel.com>
Acked-by: Martin Peres <martin.peres@linux.intel.com>