tests/perf: dramatically loosen expectations on oa-exponents
This test really quite flaky because it expects HW to emit reports
following pretty much exactly the specified timer. And there are lots
of reasons the HW might not be able to meet the timer deadline. The
design is such that it's expected to not always meet the timer.

We just want to verify that the HW was configured properly, so loosen
the checks to half the reports (of 30) meeting the timer deadline.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Acked-by: Chris Wilson <chris@chris-wilson.co.uk>