tests/intel-ci/blacklist: Blacklist kms_chamelium@hdmi-*-planes-random
Tends to generate invalid configurations using CCS planes or
configurations that exceed HW capabilities resulting in sporadic failure
with -EINVAL.

Maintaining extra constraints for Intel's HW is too costly in comparison
to the value provided.

Seems like the test is used by other vendors, so unless they are fine
with removal we'll keep it. In the meantime let's blacklist it for our
CI.

Fixes: https://gitlab.freedesktop.org/drm/intel/issues/1404
Fixes: https://gitlab.freedesktop.org/drm/intel/issues/239
Cc: Petri Latvala <petri.latvala@intel.com>
Cc: Martin Peres <martin.peres@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>