i915/gem_ctx_shared: Use a fence to group execution tightly
Use a fence to submit the pair of batches tightly together, and in
particular not allow the system time to retire in between. Then by
passing the fence along the inter-context chain, we can force the
explicit order of execution, which we currently lack as we are avoiding
the implicit fences.

In particular, without the explicit ordering the second batch could
execute before the first had managed to asynchronously bind the hidden
buffer.

References: https://gitlab.freedesktop.org/drm/intel/issues/1420
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>