i915/sysfs_heartbeat_interval: Tweak durations for 'mixed'
Be much more lenient in the face of scheduler woes that under duress
cannot give us the millisecond accuracy we require for distinguishing
good vs evil, every time.

In particular, we set the preempt timeout to greater than our 'good'
client so that when we do trigger the heartbeat check upon it, it can
always preempt in time to satisfy us.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Andi Shyti <andi.shyti@intel.com>