tests/kms_concurrent: Test for maximum number of planes
In some cases we ran out of bandwidth when executiong a test.
An error such as this can be seen on dmesg

"Bandwidth 16632 MB/s exceeds max available 6553 MB/s (7 active planes)"

The patch tests first, how many planes the platform can support with
the given resolution. The actual test is then performed by using the
calculated maximum number of planes instead of maximum number of planes
that the platform supports.

Signed-off-by: Mika Kahola <mika.kahola@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>