chamelium: Retry XMLRPC call when chamelond fails talking with a receiver
All the I2C errors we get from chameleond happen when it is talking the
receiver and seem to have nothing to do with the HDMI/DisplayPort
upstream.

This seems to happen quite often after a reset when the receiver is not
yet ready for action.

Let's keep on knocking for 10s before finally giving up and settling on
an error.

v2: igt_until_timeout() (Petri)

Fixes: https://gitlab.freedesktop.org/drm/intel/issues/217
Issue: https://bugs.chromium.org/p/chromium/issues/detail?id=993738
Cc: Imre Deak <imre.deak@intel.com>
Cc: Kunal Joshi <kunal1.joshi@intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>